Added GPL, default options and Test for BGLC
Added the GPL3 Licence Text and itegrated it in the source files
according to <http://www.gnu.org/licenses/gpl-howto.html> and
<http://www.gnu.org/prep/standards/standards.html#Command_002dLine-Interfaces>

Added a test if we can do compilation for the user by checking
if BGLC_9.exe is available to us and if wine is installed.

Added the options --version --help and --licence hard coded.

Version numbering is now hard coded at multiple points, maybe this
should change in the future.